+++ hooks/index.ts
@@ -1,4 +1,5 @@
 import Accordion from "./accordion";
+import Clipboard from "./clipboard";
 import Combobox from "./combobox";
 import Dialog from "./dialog";
 import Menu from "./menu";
@@ -6,10 +7,11 @@ import PinInput from "./pin-input";
 
 const Hooks = {
   Accordion,
+  Clipboard,
   Combobox,
   Dialog,
   Menu,
   PinInput,
 };
 
-export { Accordion, Combobox, Dialog, Menu, PinInput, Hooks };
+export { Accordion, Clipboard, Combobox, Dialog, Menu, PinInput, Hooks };